Fix uncrustify script
Add explicit path to the root source directory and allow invocation from
other directories. Also included qc_sqlite sources but excluded the SQLite
sources. Explicitly define the configuration in the MaxScale sources as
the one that will be used.

Changed from N invocations of uncrustify to one invocation with all files
as the program arguments. This should give a slight speedup when
formatting all sources.