MXS-1980 Check that Galera replication works
It should be possible to START SLAVE and STOP SLAVE irrespective
of which Galera node updates are mode to.

That will be the case if @@log_slave_updates is on and each node
in the Galera cluster have the same server id. Otherwise it will
fail with the current incarnation of BLR.